requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
aperturedb==0.4.44
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
bokeh==3.7.1
boto3==1.37.22
botocore==1.37.22
browser-use==0.1.40
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dask==2025.3.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
devtools==0.12.2
diskcache==5.6.3
distributed==2025.3.0
distro==1.9.0
Django==5.1.7
docker-pycreds==0.4.0
emoji==2.14.1
eval_type_backport==0.2.2
executing==2.2.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-genai==1.8.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gql==3.5.2
graphql-core==3.2.4
graphviz==0.20.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
image==1.5.33
imageio==2.37.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.0.4
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaggle==1.7.4.2
keepalive-socket==0.0.1
kiwisolver==1.4.8
langchain-anthropic==0.3.3
langchain-core==0.3.49
langchain-google-genai==2.1.2
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.19
lazy_loader==0.4
locket==1.0.0
lz4==4.4.3
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
narwhals==1.32.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.7
openai==1.69.0
opencv-python-headless==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
playwright==1.51.0
posthog==3.23.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0
pydantic_core==2.33.0
pyee==12.1.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
sentry-sdk==2.24.1
setproctitle==1.3.5
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.3.13
tiktoken==0.9.0
together==1.5.3
tokenizers==0.21.1
toolz==1.0.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.2
triton==3.2.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uuid_utils==0.10.0
wandb==0.19.8
wcwidth==0.2.13
weave==0.51.39
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
xyzservices==2025.1.0
yarl==1.18.3
zict==3.0.0
zstandard==0.23.0
